refactor(ol/stf): split tests into themed modules
