+++ crates/ol/stf/src/tests/deposit_withdrawal.rs
@@ -9,36 +9,36 @@ use crate::{BRIDGE_GATEWAY_ACCT_ID, BRIDGE_GATEWAY_ACCT_SERIAL, test_utils::*};
 
 #[test]
 fn test_snark_account_deposit_and_withdrawal() {
-    let snark_account_id = make_account_id(TEST_SNARK_ACCOUNT_ID);
+    let snark_acct_id = make_account_id(TEST_SNARK_ACCOUNT_ID);
     let deposit_amount = BitcoinAmount::from_sat(150_000_000); // 1.5 BTC; enough to cover withdrawal.
     let dest_subject = SubjectId::from([42u8; 32]);
 
     let genesis = OLStfFixture::builder();
-    let snark_serial = genesis.next_account_serial();
+    let snark_acct_serial = genesis.next_account_serial();
     let genesis_manifest =
-        make_deposit_manifest_for_account(1, 1, snark_serial, dest_subject, deposit_amount);
+        make_deposit_manifest_for_account(1, 1, snark_acct_serial, dest_subject, deposit_amount);
 
     let mut fixture = genesis
-        .with_genesis_snark_account(snark_account_id, |acct| {
+        .with_genesis_snark_account(snark_acct_id, |acct| {
             acct.with_state_root(make_state_root(1))
         })
         .with_genesis_manifest(genesis_manifest)
         .execute_genesis();
 
     assert_eq!(
-        fixture.account_balance(snark_account_id),
+        fixture.account_balance(snark_acct_id),
         deposit_amount,
         "Account balance should reflect the deposit"
     );
 
-    let snark_state_after_genesis = fixture.expect_snark_account(snark_account_id);
+    let account_state_after_genesis = fixture.expect_snark_account(snark_acct_id);
     assert_eq!(
-        snark_state_after_genesis.next_inbox_msg_idx(),
+        account_state_after_genesis.next_inbox_msg_idx(),
         0,
         "Next inbox idx should still be zero (no messages processed yet)"
     );
     assert_eq!(
-        snark_state_after_genesis.inbox_mmr().num_entries(),
+        account_state_after_genesis.inbox_mmr().num_entries(),
         1,
         "Should have 1 deposit message in inbox after genesis"
     );
@@ -53,7 +53,7 @@ fn test_snark_account_deposit_and_withdrawal() {
 
     let output = fixture
         .child_block()
-        .with_sau(snark_account_id, |sau| {
+        .with_sau(snark_acct_id, |sau| {
             sau.with_processed_messages(vec![deposit_msg], vec![deposit_msg_proof])
                 .output_message(
                     BRIDGE_GATEWAY_ACCT_ID,
@@ -66,7 +66,7 @@ fn test_snark_account_deposit_and_withdrawal() {
         .execute_with_outputs();
 
     assert_eq!(
-        fixture.account_balance(snark_account_id),
+        fixture.account_balance(snark_acct_id),
         BitcoinAmount::from_sat(50_000_000),
         "Account balance should be reduced by withdrawal amount"
     );
