+++ crates/ol/stf/src/tests/mod.rs
@@ -1,14 +1,18 @@
-//! Tests for snark account operations including verification and state transitions
+//! Unit tests for OL STF execution, verification, fixtures, and boundary behavior.
 
+mod chain;
 mod da_preseal_correctness;
 mod deposit_withdrawal;
-mod edge_cases;
 mod ee_predicate_update;
 mod inbox;
 mod ledger_references;
 mod limbo;
-mod multi_operations;
 mod partial_execution;
-mod stf;
-mod updates;
+mod sau_limits;
+mod sau_logs;
+mod sau_messages;
+mod sau_transfers;
 mod validation;
+mod verify_body;
+mod verify_header;
+mod verify_write_tracking;
