+++ crates/ol/stf/src/tests/sau_transfers.rs
@@ -1,0 +1,318 @@
+//! Tests for SAU transfer behavior.
+
+use strata_acct_types::BitcoinAmount;
+use strata_ledger_types::ISnarkAccountState;
+
+use crate::{BRIDGE_GATEWAY_ACCT_ID, errors::ExecError, test_utils::*};
+
+#[test]
+fn test_snark_update_success_with_transfer() {
+    let snark_acct_id = make_account_id(TEST_SNARK_ACCOUNT_ID);
+    let recipient_id = make_account_id(TEST_RECIPIENT_ID);
+
+    let mut fixture = OLStfFixture::builder()
+        .with_genesis_snark_account(snark_acct_id, |acct| {
+            acct.with_balance(BitcoinAmount::from_sat(100_000_000))
+        })
+        .with_genesis_empty_account(recipient_id)
+        .execute_genesis();
+
+    fixture
+        .child_block()
+        .with_sau(snark_acct_id, |sau| {
+            sau.transfer(recipient_id, BitcoinAmount::from_sat(30_000_000))
+                .with_state_root(make_state_root(2))
+                .with_proof(make_proof(1))
+        })
+        .execute();
+
+    // Verify balances
+    assert_eq!(
+        fixture.account_balance(snark_acct_id),
+        BitcoinAmount::from_sat(70_000_000),
+        "Sender account balance should be 100M - 30M"
+    );
+    // Check the seq no of the sender
+    assert_eq!(
+        *fixture.expect_snark_account(snark_acct_id).seqno().inner(),
+        1,
+        "Sender account seq no should increase"
+    );
+
+    assert_eq!(
+        fixture.account_balance(recipient_id),
+        BitcoinAmount::from_sat(30_000_000),
+        "Recipient should receive 30M"
+    );
+}
+
+#[test]
+fn test_snark_update_multiple_transfers() {
+    let snark_acct_id = make_account_id(TEST_SNARK_ACCOUNT_ID);
+    let recipient1_id = make_account_id(TEST_RECIPIENT_ID);
+    let recipient2_id = make_account_id(TEST_RECIPIENT_ID + 1);
+    let recipient3_id = make_account_id(TEST_RECIPIENT_ID + 2);
+
+    let mut fixture = OLStfFixture::builder()
+        .with_genesis_snark_account(snark_acct_id, |acct| {
+            acct.with_balance(BitcoinAmount::from_sat(100_000_000))
+        })
+        .with_genesis_empty_account(recipient1_id)
+        .with_genesis_empty_account(recipient2_id)
+        .with_genesis_empty_account(recipient3_id)
+        .execute_genesis();
+
+    fixture
+        .child_block()
+        .with_sau(snark_acct_id, |sau| {
+            sau.transfer(recipient1_id, BitcoinAmount::from_sat(30_000_000))
+                .transfer(recipient2_id, BitcoinAmount::from_sat(20_000_000))
+                .transfer(recipient3_id, BitcoinAmount::from_sat(10_000_000))
+                .with_state_root(make_state_root(2))
+        })
+        .execute();
+
+    // Verify all balances
+    assert_eq!(
+        fixture.account_balance(snark_acct_id),
+        BitcoinAmount::from_sat(40_000_000),
+        "Sender should have 100M - 60M = 40M"
+    );
+
+    assert_eq!(
+        fixture.account_balance(recipient1_id),
+        BitcoinAmount::from_sat(30_000_000),
+        "Recipient1 should receive 30M"
+    );
+
+    assert_eq!(
+        fixture.account_balance(recipient2_id),
+        BitcoinAmount::from_sat(20_000_000),
+        "Recipient2 should receive 20M"
+    );
+
+    assert_eq!(
+        fixture.account_balance(recipient3_id),
+        BitcoinAmount::from_sat(10_000_000),
+        "Recipient3 should receive 10M"
+    );
+}
+
+#[test]
+fn test_snark_update_partial_balance_multiple_outputs() {
+    let snark_acct_id = make_account_id(TEST_SNARK_ACCOUNT_ID);
+    let recipient1_id = make_account_id(TEST_RECIPIENT_ID);
+    let recipient2_id = make_account_id(TEST_RECIPIENT_ID + 1);
+
+    let mut fixture = OLStfFixture::builder()
+        .with_genesis_snark_account(snark_acct_id, |acct| {
+            acct.with_balance(BitcoinAmount::from_sat(100_000_000))
+        })
+        .with_genesis_empty_account(recipient1_id)
+        .with_genesis_empty_account(recipient2_id)
+        .execute_genesis();
+
+    let err = fixture
+        .child_block()
+        .with_sau(snark_acct_id, |sau| {
+            sau.transfer(recipient1_id, BitcoinAmount::from_sat(60_000_000))
+                .transfer(recipient2_id, BitcoinAmount::from_sat(50_000_000))
+                .with_state_root(make_state_root(2))
+        })
+        .execute_err();
+
+    match err.into_base() {
+        ExecError::BalanceUnderflow => {}
+        err => panic!("Expected BalanceUnderflow, got: {err:?}"),
+    }
+
+    // Verify no partial execution - all balances should be unchanged
+    assert_eq!(
+        fixture.account_balance(snark_acct_id),
+        BitcoinAmount::from_sat(100_000_000),
+        "Sender balance should be unchanged"
+    );
+
+    assert_eq!(
+        fixture.account_balance(recipient1_id),
+        BitcoinAmount::from_sat(0),
+        "Recipient1 should have no balance"
+    );
+
+    assert_eq!(
+        fixture.account_balance(recipient2_id),
+        BitcoinAmount::from_sat(0),
+        "Recipient2 should have no balance"
+    );
+}
+
+#[test]
+fn test_snark_update_zero_value_transfer() {
+    let snark_acct_id = make_account_id(TEST_SNARK_ACCOUNT_ID);
+    let recipient_id = make_account_id(TEST_RECIPIENT_ID);
+
+    let mut fixture = OLStfFixture::builder()
+        .with_genesis_snark_account(snark_acct_id, |acct| {
+            acct.with_balance(BitcoinAmount::from_sat(100_000_000))
+        })
+        .with_genesis_empty_account(recipient_id)
+        .execute_genesis();
+
+    fixture
+        .child_block()
+        .with_sau(snark_acct_id, |sau| {
+            sau.transfer(recipient_id, BitcoinAmount::from_sat(0))
+                .with_state_root(make_state_root(2))
+        })
+        .execute();
+
+    assert_eq!(
+        fixture.account_balance(snark_acct_id),
+        BitcoinAmount::from_sat(100_000_000),
+        "Sender balance should be unchanged"
+    );
+    assert_eq!(
+        *fixture.expect_snark_account(snark_acct_id).seqno().inner(),
+        1,
+        "Sequence number should still increment"
+    );
+    assert_eq!(
+        fixture.account_balance(recipient_id),
+        BitcoinAmount::from_sat(0),
+        "Recipient balance should remain 0"
+    );
+}
+
+#[test]
+fn test_snark_update_from_zero_balance_account() {
+    let snark_acct_id = make_account_id(TEST_SNARK_ACCOUNT_ID);
+    let recipient_id = make_account_id(TEST_RECIPIENT_ID);
+
+    let mut fixture = OLStfFixture::builder()
+        .with_genesis_snark_account(snark_acct_id, |acct| {
+            acct.with_balance(BitcoinAmount::from_sat(0))
+        })
+        .with_genesis_empty_account(recipient_id)
+        .execute_genesis();
+
+    let err = fixture
+        .child_block()
+        .with_sau(snark_acct_id, |sau| {
+            sau.transfer(recipient_id, BitcoinAmount::from_sat(1))
+                .with_state_root(make_state_root(2))
+        })
+        .execute_err();
+
+    match err.into_base() {
+        ExecError::BalanceUnderflow => {}
+        err => panic!("Expected BalanceUnderflow, got: {err:?}"),
+    }
+
+    assert_eq!(
+        *fixture.expect_snark_account(snark_acct_id).seqno().inner(),
+        0,
+        "Sequence number should not increment on failed transfer"
+    );
+
+    fixture
+        .child_block()
+        .with_sau(snark_acct_id, |sau| {
+            sau.transfer(recipient_id, BitcoinAmount::from_sat(0))
+                .with_state_root(make_state_root(2))
+        })
+        .execute();
+
+    assert_eq!(
+        *fixture.expect_snark_account(snark_acct_id).seqno().inner(),
+        1,
+        "Sequence number should increment even for zero transfer"
+    );
+    assert_eq!(
+        fixture.account_balance(snark_acct_id),
+        BitcoinAmount::from_sat(0),
+        "Balance should remain zero"
+    );
+    assert_eq!(
+        fixture.account_balance(recipient_id),
+        BitcoinAmount::from_sat(0),
+        "Recipient should have zero balance"
+    );
+
+    fixture
+        .child_block()
+        .with_sau(snark_acct_id, |sau| {
+            sau.transfer(recipient_id, BitcoinAmount::from_sat(0))
+                .transfer(snark_acct_id, BitcoinAmount::from_sat(0))
+                .output_message(BRIDGE_GATEWAY_ACCT_ID, BitcoinAmount::from_sat(0), vec![])
+                .with_state_root(make_state_root(3))
+        })
+        .execute();
+
+    assert_eq!(
+        *fixture.expect_snark_account(snark_acct_id).seqno().inner(),
+        2,
+        "Sequence number should increment to 2"
+    );
+}
+
+#[test]
+fn test_snark_update_self_transfer() {
+    let snark_acct_id = make_account_id(TEST_SNARK_ACCOUNT_ID);
+
+    let mut fixture = OLStfFixture::builder()
+        .with_genesis_snark_account(snark_acct_id, |acct| {
+            acct.with_balance(BitcoinAmount::from_sat(100_000_000))
+        })
+        .execute_genesis();
+
+    fixture
+        .child_block()
+        .with_sau(snark_acct_id, |sau| {
+            sau.transfer(snark_acct_id, BitcoinAmount::from_sat(30_000_000))
+                .with_state_root(make_state_root(2))
+        })
+        .execute();
+
+    assert_eq!(
+        fixture.account_balance(snark_acct_id),
+        BitcoinAmount::from_sat(100_000_000),
+        "Balance should be unchanged after self-transfer"
+    );
+    assert_eq!(
+        *fixture.expect_snark_account(snark_acct_id).seqno().inner(),
+        1,
+        "Sequence number should increment"
+    );
+}
+
+#[test]
+fn test_snark_update_exact_balance_transfer() {
+    let snark_acct_id = make_account_id(TEST_SNARK_ACCOUNT_ID);
+    let recipient_id = make_account_id(TEST_RECIPIENT_ID);
+
+    let mut fixture = OLStfFixture::builder()
+        .with_genesis_snark_account(snark_acct_id, |acct| {
+            acct.with_balance(BitcoinAmount::from_sat(100_000_000))
+        })
+        .with_genesis_empty_account(recipient_id)
+        .execute_genesis();
+
+    fixture
+        .child_block()
+        .with_sau(snark_acct_id, |sau| {
+            sau.transfer(recipient_id, BitcoinAmount::from_sat(100_000_000))
+                .with_state_root(make_state_root(2))
+        })
+        .execute();
+
+    assert_eq!(
+        fixture.account_balance(snark_acct_id),
+        BitcoinAmount::from_sat(0),
+        "Sender should have 0 balance"
+    );
+    assert_eq!(
+        fixture.account_balance(recipient_id),
+        BitcoinAmount::from_sat(100_000_000),
+        "Recipient should receive entire balance"
+    );
+}
